Move creating static & media dirs to config.yml
Since there is a likelyhood the static file directory is going to be
within the checkout directory, move creating it to the configuration
task list. This fixes an issue where a playbook fails with a "destination
path already exists" error when trying to clone the Django project into
the checkout directory.

Do the same thing for the media directory to keep things consistent.

Signed-off-by: Jason Rogena <jason@rogena.me>